+++ Sources/SwiftUIKit/Fonts/Font+FontRepresentable.swift
@@ -12,14 +12,14 @@ public extension Font {
 
     /// Get a dynamic version of the provided font.
     ///
-    /// This font type scales with dynamic type.
+    /// This value scales when using dynamic type.
     static func dynamic(_ font: FontRepresentable) -> Font {
         .custom(font.fontName, size: font.pointSize)
     }
 
     /// Get a fixed version of the provided font.
     ///
-    /// This font type doesn't scale with dynamic type.
+    /// This value doesn't scale when using dynamic type.
     static func fixed(_ font: FontRepresentable) -> Font {
         .init(font)
     }